🚚 Move onboarding components to onboarding dir
