+++ modules/ppcp-settings/resources/js/Components/Screens/Onboarding/Components/BcdcFlow.js
@@ -1,9 +1,11 @@
 import { __, sprintf } from '@wordpress/i18n';
 
-import BadgeBox, { BADGE_BOX_TITLE_BIG } from '../BadgeBox';
-import Separator from '../Separator';
-import OptionalPaymentMethods from '../OptionalPaymentMethods/OptionalPaymentMethods';
-import PricingTitleBadge from '../PricingTitleBadge';
+import BadgeBox, {
+	BADGE_BOX_TITLE_BIG,
+} from '../../../ReusableComponents/BadgeBox';
+import Separator from '../../../ReusableComponents/Separator';
+import PricingTitleBadge from '../../../ReusableComponents/PricingTitleBadge';
+import OptionalPaymentMethods from './OptionalPaymentMethods';
 
 const BcdcFlow = ( { isPayLater, storeCountry } ) => {
 	if ( isPayLater && storeCountry === 'US' ) {
