+++ modules/ppcp-settings/resources/js/Components/Screens/Onboarding/StepBusiness.js
@@ -1,9 +1,9 @@
 import { __ } from '@wordpress/i18n';
 
-import OnboardingHeader from './Components/OnboardingHeader';
 import SelectBoxWrapper from '../../ReusableComponents/SelectBoxWrapper';
 import SelectBox from '../../ReusableComponents/SelectBox';
 import { OnboardingHooks, BUSINESS_TYPES } from '../../../data';
+import OnboardingHeader from './Components/OnboardingHeader';
 
 const BUSINESS_RADIO_GROUP_NAME = 'business';
