+++ modules/ppcp-settings/resources/js/Components/Screens/Onboarding/StepPaymentMethods.js
@@ -1,11 +1,11 @@
 import { __ } from '@wordpress/i18n';
 
 import { CommonHooks, OnboardingHooks } from '../../../data';
-import OnboardingHeader from './Components/OnboardingHeader';
 import SelectBoxWrapper from '../../ReusableComponents/SelectBoxWrapper';
 import SelectBox from '../../ReusableComponents/SelectBox';
-import OptionalPaymentMethods from '../../ReusableComponents/OptionalPaymentMethods/OptionalPaymentMethods';
 import PricingDescription from '../../ReusableComponents/PricingDescription';
+import OnboardingHeader from './Components/OnboardingHeader';
+import OptionalPaymentMethods from './Components/OptionalPaymentMethods';
 
 const OPM_RADIO_GROUP_NAME = 'optional-payment-methods';
